feat: hash the long value
Closes #45

If the long value or the hash is corrupted then the card will report the long value data as empty and log an error.

Additionally, this commit writes the smaller of the raw long value and the gzipped value. Reading will attempt to ungzip, and will fall back to returning the data raw if it can't.