build(dev): add Docker infrastructure
This commit adds the ability to develop module-smartcards using Docker as an alternative approach. This does not replace the existing approach, just makes a new one available.

The main advantages are reproducibility and isolation of development environment, ease of installing dependencies regardless of the host system, and ease of orchestrating development processes. The main disadvantages are adding indirection to development systems and adding reliance on yet more software (Docker).